+++ libsemanage/tests/test_port.c
@@ -71,12 +71,12 @@ semanage_port_t *get_port_nth(int index) {
     result = semanage_port_list(sh, &records, &count);
 
     CU_ASSERT_FATAL(result >= 0);
-    CU_ASSERT_FATAL(count >= index + 1);
+    CU_ASSERT_FATAL(count >= (unsigned int) index + 1);
 
     port = records[index];
 
     for (unsigned int i = 0; i < count; i++) {
-        if (i != index)
+        if (i != (unsigned int) index)
             semanage_port_free(records[i]);
     }
