+++ libsemanage/tests/test_user.c
@@ -70,12 +70,12 @@ semanage_user_t *get_user_nth(int index) {
     result = semanage_user_list(sh, &records, &count);
 
     CU_ASSERT_FATAL(result >= 0);
-    CU_ASSERT_FATAL(count >= index + 1);
+    CU_ASSERT_FATAL(count >= (unsigned int) index + 1);
 
     user = records[index];
 
     for (unsigned int i = 0; i < count; i++) {
-        if (i != index)
+        if (i != (unsigned int) index)
             semanage_user_free(records[i]);
     }
